fix: clean up day 6 distance calc
As it turns out we can always do

root_1 = (root_1 + 1.0).floor()
root_2 = (root_2 - 1.0).ceil()

and this is equivalent to

root_1 = root_1.ceil() == root_1 ? root_1 + 1.0 : root_1.ceil()
root_2 = root_2.floor() == root_2 ? root_2 - 1.0 : root_2.floor()

The reason this works is because adding 1 and then flooring is the same
as taking the ceil but also conveniently gives us the correct result.
Same with subtracting one and then ceiling is the same as taking the
floor.